seperate component to wrap the picker
